apps/risk-engine/requirements.txt
kafka-python-ng==2.2.3
python-dotenv==1.0.1
psycopg2-binary